Pre-run clean-up.
