+++ gupb/controller/bandyta/__init__.py
@@ -1,8 +1,8 @@
 from .controller import Bandyta
 
 __all__ = [
-    "Bandyta",
-    "POTENTIAL_CONTROLLERS"
+    'Bandyta',
+    'POTENTIAL_CONTROLLERS'
 ]
 
 POTENTIAL_CONTROLLERS = [
