+++ gupb/controller/berserk/__init__.py
@@ -2,5 +2,4 @@ from gupb.controller.berserk.berserk import BerserkBot
 
 POTENTIAL_CONTROLLERS = [
     BerserkBot("Ragnar"),
-    BerserkBot("Bjorn")
 ]
